Throwing error for using a long as indexer for stringbuilder. Likely solution is to overload stringbuilder indexer with a long

+++ word-transformer/word-transformer/Controllers/WordTransformerController.cs
@@ -30,12 +30,12 @@ namespace word_transformer.Controllers
             while (currentStep < steps)
             {
                 //determine word to add
-                var index = Convert.ToInt32(Math.Floor(random.NextDouble() * dictionary.Keys.Count()) );
+                var index = Convert.ToInt64(Math.Floor(random.NextDouble() * dictionary.Keys.Count()) );
                 var sbWord = new StringBuilder(dictionary.First(d => d.Key == index).Value);
 
-                var charIndex = Convert.ToInt32(Math.Floor(random.NextDouble() * sbWord.Length));
+                var charIndex = Convert.ToInt64(Math.Floor(random.NextDouble() * sbWord.Length));
 
-                sbWord[charIndex] = characters[Convert.ToInt32(Math.Floor(random.NextDouble() * characters.Length))];
+                sbWord[charIndex] = characters[Convert.ToInt64(Math.Floor(random.NextDouble() * characters.Length))];
                 var newWord = sbWord.ToString();
 
                 if (!dictionary.ContainsValue(newWord))
